+++ server/src/auth.js
@@ -497,12 +497,9 @@ async function maySetAuthRoutes(app, genomes, basepath = '', _serverconfig = nul
 				// loosen the additional filter, to consider fewer tvs terms based on route-specific payloads or aggregation logic
 				authApi.mayAdjustFilter(req.query, ds)
 				if (req.query.daRequest) {
-					try {
-						req.query.daRequest.__protected__ = req.query.__protected__
-						authApi.mayAdjustFilter(req.query.daRequest, ds)
-					} finally {
-						delete req.query.daRequest.__protected__
-					}
+					req.query.daRequest.__protected__ = req.query.__protected__
+					authApi.mayAdjustFilter(req.query.daRequest, ds)
+					delete req.query.daRequest.__protected__
 				}
 
 				// this flag may be used by downstream code that does not have access to req argument or ds object
