+++ server/src/diffAnalysis.ts
@@ -1,15 +1,7 @@
 import crypto from 'crypto'
 import fs from 'fs'
 import path from 'path'
-import type {
-	DERequest,
-	DEImage,
-	ExpressionInput,
-	GeneDEEntry,
-	GenesetEnrichmentRequest,
-	GenesetEnrichmentResponse
-} from '#types'
-import { run_python } from '@sjcrh/proteinpaint-python'
+import type { DERequest, DEImage, ExpressionInput, GeneDEEntry } from '#types'
 import { run_rust } from '@sjcrh/proteinpaint-rust'
 import { run_R } from '@sjcrh/proteinpaint-r'
 import { getData } from './termdb.matrix.js'
@@ -21,7 +13,7 @@ import { formatElapsedTime } from '#shared'
 
 // Any drift in this function between the volcano path and the GSEA path will
 // cause cache-miss storms. Keep it minimal and purely structural.
-function stableStringify(v: any): string {
+export function stableStringify(v: any): string {
 	if (v === null || typeof v !== 'object') return JSON.stringify(v) ?? 'null'
 	if (Array.isArray(v)) return '[' + v.map(stableStringify).join(',') + ']'
 	const keys = Object.keys(v).sort()
@@ -533,176 +525,4 @@ async function doReadOrRecompute(cacheId: string, daRequest: DERequest, genomes:
 		images: fresh.images,
 		bcv: fresh.bcv
 	}
-}
-
-// -----------------------------------------------------------------------------
-// GSEA (blitzgsea) — deterministic pickle caching
-// -----------------------------------------------------------------------------
-
-/** Resolve the `genes` + `fold_change` inputs that every GSEA path needs.
- * Shared between `computeGSEA` (blitzgsea), the GSEA route's cerno
- * branch, and the `fetchDE` short-circuit — a single source of truth for
- * "where do the arrays come from" so the cacheId/daRequest contract is
- * enforced in one place. */
-export async function resolveGseaGenesAndFoldChange({
-	q,
-	genomes
-}: {
-	q: GenesetEnrichmentRequest
-	genomes: any
-}): Promise<{ genes: string[]; fold_change: number[] }> {
-	if (q.cacheId) {
-		// A cacheId without daRequest is malformed — without the DE
-		// snapshot there's no way to verify the id or recompute on a cache
-		// miss. The 'daCacheMissing' literal matches the client's sayerror
-		// regex so stale sessions still get the reopen-the-volcano
-		// message.
-		if (!q.daRequest) throw new Error('daCacheMissing')
-		const result = await readCacheFileOrRecompute({ daRequest: q.daRequest as DERequest, genomes })
-		if (result.cacheId !== q.cacheId) throw new Error('cacheId does not match daRequest')
-		return {
-			genes: result.geneData.map(g => g.gene_name),
-			fold_change: result.geneData.map(g => g.fold_change)
-		}
-	}
-	// Inline path (legacy single-cell). Reject early so we don't pass
-	// undefined down to Python/Rust.
-	if (!q.genes || !q.fold_change) throw new Error('requires genes and fold_change when cacheId is absent')
-	return { genes: q.genes, fold_change: q.fold_change }
-}
-
-/** Compute the deterministic pickle filename for a blitzgsea run. Hash
- * includes everything that affects the blitzgsea result: the signature
- * arrays (genes + fold_change), the gene-set library selector, the
- * permutation count, and the coding-genes filter flag. Method is always
- * 'blitzgsea' for this helper; included in the hash for future-proofing. */
-export function computeGseaPickleId({
-	genes,
-	fold_change,
-	geneSetGroup,
-	num_permutations,
-	filter_non_coding_genes
-}: {
-	genes: string[]
-	fold_change: number[]
-	geneSetGroup: string
-	num_permutations: number | undefined
-	filter_non_coding_genes: boolean
-}): string {
-	const keyInputs = {
-		genes,
-		fold_change,
-		geneSetGroup,
-		num_permutations: num_permutations ?? null,
-		filter_non_coding_genes,
-		method: 'blitzgsea'
-	}
-	const key = stableStringify(keyInputs)
-	const hash = crypto.createHash('sha256').update(key).digest('hex').slice(0, 32)
-	return `gsea_${hash}.pkl`
-}
-
-/** In-flight blitzgsea computations keyed by pickle filename. Dedupes
- * concurrent callers with identical inputs so only one Python invocation
- * runs per unique pickleId. Same pattern as `pendingReadOrRecompute`. Only
- * gates the initial-computation path (no geneset_name); detail-plot
- * requests bypass this map. */
-const pendingGseaCompute = new Map<string, Promise<GenesetEnrichmentResponse | string>>()
-
-/** Single entry point for GSEA (blitzgsea) — resolves genes+fold_change,
- * computes a deterministic pickle filename, calls Python which either
- * loads the pickle or computes and writes it, and returns the result.
- *
- * For cacheId-bearing requests, genes+fold_change are pulled from the DA
- * cache via `readCacheFileOrRecompute`. For the legacy single-cell path,
- * they come inline on the request. */
-export async function computeGSEA({
-	q,
-	genomes
-}: {
-	q: GenesetEnrichmentRequest
-	genomes: any
-}): Promise<GenesetEnrichmentResponse | string> {
-	const { genes, fold_change } = await resolveGseaGenesAndFoldChange({ q, genomes })
-
-	// Always recompute the pickle filename from the GSEA inputs. The client
-	// no longer carries `pickle_file` — the detail-plot request carries the
-	// same inputs as the initial enrichment, so the hash resolves to the
-	// same filename and Python finds the previously-written pickle.
-	const pickle_file = computeGseaPickleId({
-		genes,
-		fold_change,
-		geneSetGroup: q.geneSetGroup,
-		num_permutations: q.num_permutations,
-		filter_non_coding_genes: q.filter_non_coding_genes
-	})
-
-	// In-flight dedup for the initial computation only (detail-plot is a
-	// different unit of work per geneset_name and fast).
-	if (!q.geneset_name) {
-		const inFlight = pendingGseaCompute.get(pickle_file)
-		if (inFlight) return inFlight
-	}
-
-	const work = runGseaPython({ q, genomes, genes, fold_change, pickle_file })
-
-	if (!q.geneset_name) {
-		pendingGseaCompute.set(pickle_file, work)
-		return work.finally(() => {
-			pendingGseaCompute.delete(pickle_file)
-		})
-	}
-
-	return work
-}
-
-async function runGseaPython({
-	q,
-	genomes,
-	genes,
-	fold_change,
-	pickle_file
-}: {
-	q: GenesetEnrichmentRequest
-	genomes: any
-	genes: string[]
-	fold_change: number[]
-	pickle_file: string
-}): Promise<GenesetEnrichmentResponse | string> {
-	const cachedir_gsea = path.join(serverconfig.cachedir, 'gsea')
-	const genesetenrichment_input: any = {
-		genes,
-		fold_change,
-		db: genomes[q.genome].termdbs.msigdb.cohort.db.connection.name,
-		geneset_group: q.geneSetGroup,
-		genedb: path.join(serverconfig.tpmasterdir, genomes[q.genome].genedb.dbfile),
-		filter_non_coding_genes: q.filter_non_coding_genes,
-		cachedir: cachedir_gsea,
-		pickle_file,
-		geneset_name: q.geneset_name,
-		num_permutations: q.num_permutations
-	}
-
-	const time1 = new Date().valueOf()
-	const gsea_output: string = await run_python('gsea.py', '/' + JSON.stringify(genesetenrichment_input))
-	mayLog('Time taken to run blitzgsea:', formatElapsedTime(Date.now() - time1))
-
-	let result: any
-	let data_found = false
-	let image_found = false
-	for (const line of gsea_output.split('\n')) {
-		if (line.startsWith('result: ')) {
-			result = JSON.parse(line.replace('result: ', ''))
-			data_found = true
-		} else if (line.startsWith('image: ')) {
-			result = JSON.parse(line.replace('image: ', ''))
-			image_found = true
-		} else {
-			mayLog(line)
-		}
-	}
-
-	if (data_found) return result as GenesetEnrichmentResponse
-	if (image_found) return path.join(cachedir_gsea, result.image_file)
-	throw new Error('data or image not found in gsea output; this should not happen')
 }
